+++ apps/matching-engine/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "matching-engine"
+name = "smolexchange"
 version = "0.1.0"
 edition = "2021"
